+++ src/Neo.Compiler.CSharp/MethodConvert/SourceConvert.cs
@@ -1,0 +1,57 @@
+// Copyright (C) 2015-2023 The Neo Project.
+//
+// The Neo.Compiler.CSharp is free software distributed under the MIT
+// software license, see the accompanying file LICENSE in the main directory
+// of the project or http://www.opensource.org/licenses/mit-license.php
+// for more details.
+//
+// Redistribution and use in source and binary forms with or without
+// modifications are permitted.
+
+extern alias scfx;
+
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+
+namespace Neo.Compiler;
+
+partial class MethodConvert
+{
+    private void ConvertSource(SemanticModel model)
+    {
+        if (SyntaxNode is null) return;
+        for (byte i = 0; i < Symbol.Parameters.Length; i++)
+        {
+            IParameterSymbol parameter = Symbol.Parameters[i].OriginalDefinition;
+            byte index = i;
+            if (!Symbol.IsStatic) index++;
+            _parameters.Add(parameter, index);
+        }
+        switch (SyntaxNode)
+        {
+            case AccessorDeclarationSyntax syntax:
+                if (syntax.Body is not null)
+                    ConvertStatement(model, syntax.Body);
+                else if (syntax.ExpressionBody is not null)
+                    using (InsertSequencePoint(syntax.ExpressionBody.Expression))
+                        ConvertExpression(model, syntax.ExpressionBody.Expression);
+                else
+                    ConvertNoBody(syntax);
+                break;
+            case ArrowExpressionClauseSyntax syntax:
+                using (InsertSequencePoint(syntax))
+                    ConvertExpression(model, syntax.Expression);
+                break;
+            case BaseMethodDeclarationSyntax syntax:
+                if (syntax.Body is null)
+                    using (InsertSequencePoint(syntax.ExpressionBody!.Expression))
+                        ConvertExpression(model, syntax.ExpressionBody.Expression);
+                else
+                    ConvertStatement(model, syntax.Body);
+                break;
+            default:
+                throw new CompilationException(SyntaxNode, DiagnosticId.SyntaxNotSupported, $"Unsupported method body:{SyntaxNode}");
+        }
+        _initslot = !_inline;
+    }
+}
